profile statistics logic done
